requirements.txt
openai>=1.14.0
pydantic>=2.0.0
duckduckgo-search>=7.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
